% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_column_names}
\alias{clean_column_names}
\title{Clean column names using \code{janitor::clean_names}}
\usage{
clean_column_names(df)
}
\arguments{
\item{df}{A data frame with column names to clean.}
}
\value{
A data frame with cleaned column names.
}
\description{
Converts all column names in a data frame to consistent \code{snake_case} format.
}
\examples{
clean_column_names(data.frame("My Col" = 1:3))
}
