% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{compute_avg_amounts}
\alias{compute_avg_amounts}
\title{Compute row-wise averages for \code{BILL_AMT} and \code{PAY_AMT} columns}
\usage{
compute_avg_amounts(df)
}
\arguments{
\item{df}{A data frame containing columns like \code{bill_amt1} to \code{bill_amt6},
and \code{pay_amt1} to \code{pay_amt6}.}
}
\value{
A modified data frame with average columns and without the original billing/payment columns.
}
\description{
Adds two new columns: \code{avg_bill_amt} and \code{avg_pay_amt} to the data frame,
and removes the original billing and payment columns.
}
