% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_target_plot.R
\name{make_target_plot}
\alias{make_target_plot}
\title{Create a bar plot of the default payment distribution}
\usage{
make_target_plot(df)
}
\arguments{
\item{df}{A data frame containing a column named \code{default_payment_next_month}.}
}
\value{
A \code{ggplot} object showing the bar plot of default status.
}
\description{
This function takes a data frame that contains the column \code{default_payment_next_month}
and produces a bar plot showing the distribution of default status.
}
\examples{
make_target_plot(df)  # df must include the column `default_payment_next_month`
}
