% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_abstraction.R
\name{prepare_knn_data}
\alias{prepare_knn_data}
\title{Prepare normalized predictors and extract labels}
\usage{
prepare_knn_data(df_train, df_test, target = "default_payment_next_month")
}
\arguments{
\item{df_train}{Training data}

\item{df_test}{Testing data}

\item{target}{Column name of target variable}
}
\value{
A list with normalized train/test data and labels
}
\description{
Prepare normalized predictors and extract labels
}
\examples{
df <- data.frame(
  feature1 = c(1, 2, 3, 4, 5),
  feature2 = c(5, 4, 3, 2, 1),
  default_payment_next_month = c(0, 1, 0, 1, 0)
)
split <- split_dataset(df)
prepare_knn_data(split$train, split$test, "default_payment_next_month")
}
