% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling_abstraction.R
\name{save_accuracy_plot}
\alias{save_accuracy_plot}
\title{Save accuracy vs. k plot to a file}
\usage{
save_accuracy_plot(accuracy_df, out_path)
}
\arguments{
\item{accuracy_df}{Data frame with k and accuracy}

\item{out_path}{Output file path for the plot (PNG)}
}
\description{
Creates a line plot from KNN accuracy results and saves it as a PNG.
}
\examples{
\dontrun{
df <- data.frame(k = 1:5, accuracy = c(0.6, 0.7, 0.75, 0.72, 0.8))
save_accuracy_plot(df, "knn_accuracy.png")
}

}
