% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{split_data}
\alias{split_data}
\title{Split a data frame into training and testing sets}
\usage{
split_data(df, train_size = 0.8)
}
\arguments{
\item{df}{A data frame to split.}

\item{train_size}{Proportion of data to use for training (default is 0.8).}
}
\value{
A list with two data frames: \code{train} and \code{test}.
}
\description{
Randomly splits a data frame into a training set and a test set based on the given proportion.
}
